requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-guardian==2.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
drf-yasg==1.21.6
inflection==0.5.1
kombu==5.3.7
packaging==24.0
prompt-toolkit==3.0.43
psycopg2==2.9.6
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
coreapi==2.3.3
isort==5.12.0
black==23.3.0